ai/pyproject.toml
[project]
name = "ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-mistralai>=0.2.10",
    "langchain-pinecone>=0.2.8",
    "motor>=3.7.1",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.3",
]
